Returning back functionality from Favourite tracks to Mainlist.tsx 😀
